fixes to several boards and soft I2C
- fixed I2C on AVR
- fixed I2C on SAMD21
- added alternative I2C via Arduino Wire to LPC176x
-fixed soft I2C

+++ uCNC/src/modules/softi2c.c
@@ -65,7 +65,7 @@ static uint8_t softi2c_write(softi2c_port_t *port, uint8_t c, bool send_start, b
 	{
 		// init
 		port->sda(true);
-		uint8_t timeout = ms_timeout;
+		uint32_t timeout = ms_timeout;
 		__TIMEOUT_MS__(timeout)
 		{
 			if (port->get_sda())
